deploy/requirements.txt
a2a-sdk==0.2.16
absolufy-imports==0.3.1
agent-starter-pack==0.11.2
annotated-types==0.7.0
anyio==4.10.0
arrow==1.3.0
attrs==25.3.0
Authlib==1.6.1
backoff==2.2.1
beautifulsoup4==4.13.4
binaryornot==0.4.4
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
contourpy==1.3.3
cookiecutter==2.6.0
cryptography==45.0.5
cssselect2==0.8.0
cycler==0.12.1
cyclopts==3.22.5
diagrams==0.24.4
distlib==0.4.0
dnspython==2.7.0
docstring_parser==0.17.0
docutils==0.22
docx2pdf==0.1.8
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastmcp==2.11.2
filelock==3.18.0
fire==0.7.0
fonttools==4.59.0
google-adk==1.9.0
google-api-core==2.25.1
google-api-python-client==2.178.0
google-assistant-sdk==0.6.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-access-context-manager==0.2.2
google-cloud-aiplatform==1.106.0
google-cloud-appengine-logging==1.6.2
google-cloud-asset==3.30.1
google-cloud-audit-log==0.3.2
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-logging==3.12.1
google-cloud-org-policy==1.14.0
google-cloud-os-config==1.20.2
google-cloud-resource-manager==1.14.2
google-cloud-secret-manager==2.24.0
google-cloud-speech==2.33.0
google-cloud-storage==2.19.0
google-cloud-trace==1.16.2
google-crc32c==1.7.1
google-genai==1.28.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphviz==0.20.3
greenlet==3.2.3
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lazy-object-proxy==1.11.0
lxml==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mcp==1.12.3
mdurl==0.1.2
more-itertools==10.7.0
nodeenv==1.9.1
numpy==2.2.6
oauthlib==3.3.1
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-gcp-trace==1.9.0
opentelemetry-resourcedetector-gcp==1.9.0a0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
packaging==25.0
pandas==2.3.1
parse==1.20.2
pathable==0.4.4
pdf2docx==0.5.8
pillow==11.3.0
platformdirs==4.3.8
pre_commit==4.2.0
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydyf==0.11.0
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pyperclip==1.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.3
requests==2.32.4
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.26.0
rsa==4.9.1
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.42
sse-starlette==3.0.2
starlette==0.47.2
tenacity==8.5.0
termcolor==3.1.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typer==0.16.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.33.1
watchdog==6.0.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.1
zipp==3.23.0
zopfli==0.2.3.post1
google-adk[a2a]
python-dotenv
uvicorn
uv
python-docx
pandas
openpyxl
reportlab
pdf2docx
docx2pdf
google-cloud-asset
google-assistant-sdk[cli]